Update o/p directory from 'client' to 'gluonfx'. Also rename tasks.
