+++ src/main/java/com/gluonhq/gradle/tasks/NativeBuildTask.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2020, Gluon
+ * Copyright (c) 2019, 2021, Gluon
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,28 +29,21 @@
  */
 package com.gluonhq.gradle.tasks;
 
-import javax.inject.Inject;
-
+import org.gradle.api.DefaultTask;
 import org.gradle.api.Project;
 import org.gradle.api.tasks.TaskAction;
 
-import com.gluonhq.substrate.SubstrateDispatcher;
+import javax.inject.Inject;
 
-public class ClientNativeRun extends ClientNativeBase {
+public class NativeBuildTask extends DefaultTask {
+
     @Inject
-    public ClientNativeRun(Project project) {
-        super(project);
+    public NativeBuildTask(Project project) {
+        dependsOn(project.getTasks().findByName("nativeCompile"), project.getTasks().findByName("nativeLink"));
     }
 
     @TaskAction
     public void action() {
-        getProject().getLogger().info("ClientNativeRun action");
-
-        try {
-            SubstrateDispatcher dispatcher = new ConfigBuild(project).createSubstrateDispatcher();
-            dispatcher.nativeRun();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        getProject().getLogger().info("ClientNativeBuild action");
     }
 }
