+++ src/main/java/com/gluonhq/gradle/tasks/NativeInstallTask.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Gluon
+ * Copyright (c) 2019, 2021, Gluon
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,21 +29,34 @@
  */
 package com.gluonhq.gradle.tasks;
 
-import org.gradle.api.DefaultTask;
-import org.gradle.api.Project;
-import org.gradle.api.plugins.JavaPlugin;
-
 import javax.inject.Inject;
 
-abstract class ClientNativeBase extends DefaultTask {
+import org.gradle.api.GradleException;
+import org.gradle.api.Project;
+import org.gradle.api.tasks.TaskAction;
 
-    final Project project;
+import com.gluonhq.substrate.SubstrateDispatcher;
 
+public class NativeInstallTask extends NativeBaseTask {
     @Inject
-    public ClientNativeBase(Project project) {
-        this.project = project;
-        project.getPluginManager().withPlugin("java", e ->
-                dependsOn(project.getTasks().findByName(JavaPlugin.CLASSES_TASK_NAME),
-                          project.getTasks().findByName(JavaPlugin.PROCESS_RESOURCES_TASK_NAME)));
+    public NativeInstallTask(Project project) {
+        super(project);
+    }
+
+    @TaskAction
+    public void action() {
+        getProject().getLogger().info("ClientNativeInstall action");
+
+        boolean result;
+        try {
+            SubstrateDispatcher dispatcher = new ConfigBuild(project).createSubstrateDispatcher();
+            result = dispatcher.nativeInstall();
+        } catch (Exception e) {
+            throw new GradleException("Failed to install", e);
+        }
+
+        if (!result) {
+            throw new GradleException("Installing failed");
+        }
     }
 }
